Consolidate IBM Cloud logo
There are now two versions of the logo: one that can be edited in a
graphics editor ("editable" suffix) and one that can be displayed in a
web browser ("browser" suffix).

The reason for this is that the official SVG version of the logo (e.g. on
https://identity-2.eu-central.iam.cloud.ibm.com/ui/interaction_error.jsp)
includes a mask to achieve the shape and the gradient of the icon. This
can be displayed in a browser, but not edited in a graphics editor.

To create editable versions, a PNG-to-SVG converted icon has been taken
from https://getlogovector.com/download-logo-vector-480e# and extended
with the text to achieve the three versions of the editable logo (icon,
horizontal, and stacked). The PNG-to-SVG converted icon is naturally not
clean, thus it should be used only in small sizes.

---
sources:
  - https://identity-2.eu-central.iam.cloud.ibm.com/ui/interaction_error.jsp (horizontal/browser)
  - https://www.vectorlogo.zone/logos/ibm_cloud/index.html (icon/browser)
  - https://getlogovector.com/download-logo-vector-480e# (icon/editable)
note: >
  Editable versions have been created by extending editable icon with
  text from browser versions.