Add YAML meta file in root directory with meta info about logos
The meta information includes the full name of the company/project of
the logo, any aliases of this name, and the ID (which corresponds to
the directory in which the logo files are saved).

At a later point, tags could be added to this meta information about
each company/project, and a new YAML document listing all valid tags
could be added to the meta file.

The goal of this is to cater for a future web frontend for the logo
collection.